fix(ecs-browser): fix browser scrolling when there are a lot of entities
we need to decide on how we want to structure the UI components. right now, it is difficult to place
components on the screen because the component renderer is wrapping them in its own divs. we should
have the position it lives on the screen be defined as part of registerUiComponent