feat(relay): check account balance is greater than a configurable amount, rather than just >0
Right now, we relay messages for accouts with any balance. This fails to sybil-protect the relay
system as an attacker could send 1 wei to hundreds of accounts. This commit adds a field to the
relay server configuration specifying the minimum balance an account must have to get its messages
relayed.