feat: erc1155-puppet world-module
building a rough-on-the-edge 1st iteration of what it would look like to have an ERC1155 world-module.
Respects OZ/Enjin standards, with only small deviations from it so far (mostly massaging the ERC to make it fit into a MUD context)

Please disregard ERC1155/System.sol for now, as it's just there as a filler for now.
ERC1155Module.sol is also highly susceptible to change in later commits.

No unit test files are present as I don't want to break the CI/CD pipeline just yet.