Fix: ERC1155.t.sol
a few assumptions were wrong for some of the fuzzed unit tests, so I added extra checks to cover for those edge cases

Also had to make the test contract itself implement the IERC1155Receiver interface since there were unit tests checking wat would happen for transferFrom-itself, and to do that it starts by minting tokens to itself.
Because minting is actually a safe transfer to that address, and the test contract is obviously not an EOA, it means that the minting subroutine will fail if it's not ERC1155receiver-compliant.

Anyhow, everything works now :thumbsup: